requirements.txt
python-dotenv
breeze-connect
kiteconnect
psycopg2-binary